Avoid circular imports from Assignment and CaretValue rules
Importing entire directories instead of individual files lead to cases
where files imports would form a circle. Although this wasn't a problem
for the compiler, it presented a problem for automated testing. By using
more specific import statements, these import cycles are removed.

To support more specific imports, some functions and classes are
reorganized to be in their own file rather than in a general-purpose
file containing commonly used functions and classes.

Add the new HasFlags mixin for use by FlagRule and AddElementRule.
Add support for bigint and FshCanonical values when producing FSH for an
AssignmentRule or CaretValueRule.
Fix bug with CardRule producing incorrect FSH when min == 0.
