+++ src/fhirtypes/CodeSystem.ts
@@ -3,8 +3,8 @@ import { Meta } from './specialTypes';
 import { Extension } from '../fshtypes';
 import { Narrative, Resource, Identifier, CodeableConcept, Coding } from './dataTypes';
 import { ContactDetail, UsageContext } from './metaDataTypes';
-import { HasName, HasId } from './common';
-import { applyMixins } from '../utils';
+import { HasName, HasId } from './mixins';
+import { applyMixins } from '../utils/Mixin';
 
 /**
  * Class representing a FHIR R4 CodeSystem
