+++ src/fhirtypes/StructureDefinition.ts
@@ -15,13 +15,13 @@ import {
 import {
   getArrayIndex,
   setPropertyOnDefinitionInstance,
-  HasName,
-  HasId,
   isInheritedResource,
   isExtension
 } from './common';
+import { HasName, HasId } from './mixins';
 import { Fishable, Type } from '../utils/Fishable';
-import { applyMixins, parseFSHPath, assembleFSHPath } from '../utils';
+import { applyMixins } from '../utils/Mixin';
+import { parseFSHPath, assembleFSHPath } from '../utils/PathUtils';
 import { InstanceDefinition } from './InstanceDefinition';
 import { isUri } from 'valid-url';
