+++ src/fshtypes/InstanceUsage.ts
@@ -1,0 +1,5 @@
+export type InstanceUsage = 'Example' | 'Definition' | 'Inline';
+
+export function isInstanceUsage(s: string): s is InstanceUsage {
+  return ['Example', 'Definition', 'Inline'].indexOf(s) >= 0;
+}
