+++ src/fshtypes/rules/CaretValueRule.ts
@@ -1,7 +1,10 @@
 import { Rule } from './Rule';
 import { AssignmentValueType } from './AssignmentRule';
-import { FshCanonical, FshCode, FshQuantity, FshRatio, FshReference } from '..';
-import { InstanceDefinition } from '../../fhirtypes';
+import { FshCanonical } from '../FshCanonical';
+import { FshCode } from '../FshCode';
+import { FshQuantity } from '../FshQuantity';
+import { FshRatio } from '../FshRatio';
+import { FshReference } from '../FshReference';
 import { EOL } from 'os';
 
 export class CaretValueRule extends Rule {
@@ -27,12 +30,16 @@ export class CaretValueRule extends Rule {
       this.value instanceof FshReference
     ) {
       value = this.value.toString();
-    } else if (this.value instanceof InstanceDefinition) {
-      value = this.value._instanceMeta.name;
-    } else if (typeof this.value === 'boolean' || typeof this.value === 'number') {
+    } else if (
+      typeof this.value === 'boolean' ||
+      typeof this.value === 'number' ||
+      typeof this.value === 'bigint'
+    ) {
       value = this.value;
     } else if (typeof this.value === 'string') {
       value = this.isInstance ? this.value : `"${this.value}"`;
+    } else if (this.value) {
+      value = this.value._instanceMeta.name;
     }
     const lines: string[] = [];
     lines.push(`* ${this.path !== '' ? this.path + ' ' : ''}^${this.caretPath} = ${value}`);
